clinical_dataflow_optimizer/test_longcat_focused.py
"""
Focused test for LongCat AI integration components
Tests API connectivity, client functionality, and agent integration
"""

import sys
import os
from pathlib import Path

# Add project root to path
PROJECT_ROOT = Path(__file__).parent
sys.path.insert(0, str(PROJECT_ROOT))

# Load environment variables
try:
    from dotenv import load_dotenv
    load_dotenv()
except ImportError:
    # dotenv not available, rely on system environment
    pass

from core.longcat_integration import longcat_client, LongCatClient
from config.settings import DEFAULT_LONGCAT_CONFIG
from agents.agent_framework import ReconciliationAgent

def test_env_loading():
    """Test environment variable loading"""
    print("🔧 Testing environment variable loading...")

    api_key = os.getenv('API_KEY_Longcat')
    if api_key:
        print(f"✅ API key loaded: {api_key[:10]}...")
        return True
    else:
        print("❌ API key not found in environment")
        return False

def test_config_loading():
    """Test configuration loading"""
    print("\n⚙️ Testing configuration loading...")

    config = DEFAULT_LONGCAT_CONFIG
    print(f"API Key from config: {config.api_key[:10] if config.api_key else 'None'}...")
    print(f"Base URL: {config.base_url}")
    print(f"Model: {config.model}")
    print(f"Thinking Model: {config.thinking_model}")

    if config.api_key:
        print("✅ Configuration loaded successfully")
        return True
    else:
        print("❌ Configuration missing API key")
        return False

def test_client_initialization():
    """Test LongCat client initialization"""
    print("\n🔌 Testing LongCat client initialization...")

    try:
        client = LongCatClient()
        print("✅ Client initialized successfully")
        print(f"Session headers: Authorization present = {'Authorization' in client.session.headers}")
        return True
    except Exception as e:
        print(f"❌ Client initialization failed: {e}")
        return False

def test_basic_api_connectivity():
    """Test basic API connectivity"""
    print("\n🌐 Testing basic API connectivity...")

    try:
        # Test with a simple message
        messages = [
            {"role": "system", "content": "You are a helpful assistant."},
            {"role": "user", "content": "Hello! Please respond with just 'Hello from LongCat'"}
        ]

        response = longcat_client.chat_completion(messages, max_tokens=10)

        if 'choices' in response and len(response['choices']) > 0:
            content = response['choices'][0]['message']['content']
            print(f"✅ API response received: {content}")
            print(f"Model used: {response.get('model', 'unknown')}")
            print(f"Tokens used: {response.get('usage', {}).get('total_tokens', 'unknown')}")
            return True
        else:
            print(f"❌ Unexpected response format: {response}")
            return False

    except Exception as e:
        print(f"❌ API connectivity test failed: {e}")
        return False

def test_agent_integration():
    """Test agent integration with LongCat"""
    print("\n🤖 Testing agent integration...")

    try:
        # Create an agent
        agent = ReconciliationAgent("TestRex")

        # Test the enhanced reasoning method
        context = "Patient has missing AE form for reported SAE"
        task = "Generate recommendation for site query"
        data = {
            "subject_id": "TEST-001",
            "sae_date": "2024-01-15",
            "missing_form": "AE_Form"
        }

        # Test if the method exists
        if hasattr(agent, 'enhance_reasoning_with_longcat'):
            print("✅ Agent has LongCat reasoning method")

            # Try to call it (will fail gracefully if API doesn't work)
            try:
                reasoning = agent.enhance_reasoning_with_longcat(context, task, data)
                print(f"✅ Enhanced reasoning generated: {reasoning[:100]}...")
                return True
            except Exception as e:
                print(f"⚠️ Enhanced reasoning failed (expected if API down): {e}")
                return True  # Method exists, API failure is separate
        else:
            print("❌ Agent missing LongCat reasoning method")
            return False

    except Exception as e:
        print(f"❌ Agent integration test failed: {e}")
        return False

def test_narrative_generation():
    """Test narrative generation capability"""
    print("\n📝 Testing narrative generation...")

    try:
        patient_data = {"subject_id": "TEST-001", "issues": ["Missing visit", "Open query"]}
        issues = ["Missing scheduled visit", "Unanswered data query"]
        recommendations = ["Contact site", "Follow up on query"]

        narrative = longcat_client.generate_narrative(patient_data, issues, recommendations)
        print(f"✅ Narrative generated: {narrative[:150]}...")
        return True

    except Exception as e:
        print(f"❌ Narrative generation failed: {e}")
        return False

def test_anomaly_explanation():
    """Test anomaly explanation capability"""
    print("\n🔍 Testing anomaly explanation...")

    try:
        anomaly_data = {
            "type": "missing_visit",
            "subject_id": "TEST-001",
            "days_overdue": 30
        }
        context = "Patient has missed a critical study visit"

        explanation = longcat_client.explain_anomaly(anomaly_data, context)
        print(f"✅ Anomaly explanation generated: {explanation[:150]}...")
        return True

    except Exception as e:
        print(f"❌ Anomaly explanation failed: {e}")
        return False

def main():
    """Run all LongCat integration tests"""
    print("🚀 LongCat AI Integration Test Suite")
    print("=" * 50)

    tests = [
        ("Environment Loading", test_env_loading),
        ("Configuration Loading", test_config_loading),
        ("Client Initialization", test_client_initialization),
        ("API Connectivity", test_basic_api_connectivity),
        ("Agent Integration", test_agent_integration),
        ("Narrative Generation", test_narrative_generation),
        ("Anomaly Explanation", test_anomaly_explanation),
    ]

    results = []
    for test_name, test_func in tests:
        print(f"\n{'='*20} {test_name} {'='*20}")
        result = test_func()
        results.append((test_name, result))

    # Summary
    print("\n" + "=" * 50)
    print("📊 TEST RESULTS SUMMARY")
    print("=" * 50)

    passed = 0
    total = len(results)

    for test_name, result in results:
        status = "✅ PASS" if result else "❌ FAIL"
        print(f"{status} {test_name}")
        if result:
            passed += 1

    print(f"\nOverall: {passed}/{total} tests passed")

    if passed >= 4:  # Core integration working
        print("🎉 LongCat AI integration is WORKING!")
        print("   - Core components initialized")
        print("   - Agent integration functional")
        print("   - API client ready (may need valid API key for full functionality)")
    else:
        print("⚠️ LongCat AI integration has issues that need attention")

    return passed == total

if __name__ == "__main__":
    success = main()
    sys.exit(0 if success else 1)