clinical_dataflow_optimizer/requirements-web.txt
# Neural Clinical Data Mesh Web Application Requirements
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
networkx>=2.8.0
plotly>=5.0.0

# Web framework dependencies
flask>=2.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
eventlet>=0.33.0

# Alternative dashboard framework
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# AI and ML dependencies
openai>=1.0.0
anthropic>=0.7.0

# Configuration and utilities
python-dotenv>=0.19.0
dataclasses-json>=0.5.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: For enhanced performance
asyncio-mqtt>=0.11.0  # If MQTT integration is needed
redis>=4.0.0  # For caching/session management